@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.4.4",
+  "version": "8.5.1",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -21,17 +21,17 @@
   "dependencies": {
     "bson": "^6.7.0",
     "kareem": "2.6.3",
-    "mongodb": "6.6.2",
+    "mongodb": "6.7.0",
     "mpath": "0.9.0",
     "mquery": "5.0.0",
     "ms": "2.1.3",
     "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.24.6",
-    "@babel/preset-env": "7.24.6",
-    "@typescript-eslint/eslint-plugin": "^6.2.1",
-    "@typescript-eslint/parser": "^6.2.1",
+    "@babel/core": "7.24.7",
+    "@babel/preset-env": "7.24.7",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
     "acquit": "1.3.0",
     "acquit-ignore": "0.2.1",
     "acquit-require": "0.1.1",
@@ -47,27 +47,27 @@
     "eslint": "8.57.0",
     "eslint-plugin-markdown": "^5.0.0",
     "eslint-plugin-mocha-no-only": "1.2.0",
-    "express": "^4.18.1",
+    "express": "^4.19.2",
     "fs-extra": "~11.2.0",
-    "highlight.js": "11.8.0",
+    "highlight.js": "11.9.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
     "markdownlint-cli2": "^0.13.0",
     "marked": "4.3.0",
     "mkdirp": "^3.0.1",
-    "mocha": "10.4.0",
-    "moment": "2.x",
-    "mongodb-memory-server": "9.3.0",
+    "mocha": "10.6.0",
+    "moment": "2.30.1",
+    "mongodb-memory-server": "9.4.0",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.3",
     "q": "1.5.1",
     "sinon": "18.0.0",
     "stream-browserify": "3.0.0",
-    "tsd": "0.31.0",
-    "typescript": "5.4.5",
-    "uuid": "9.0.1",
-    "webpack": "5.91.0"
+    "tsd": "0.31.1",
+    "typescript": "5.5.3",
+    "uuid": "10.0.0",
+    "webpack": "5.92.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
