@@ -1,29 +1,36 @@
-import express from "express"
-import dotenv from "dotenv"
-import mongoose from "mongoose"
+import express from "express";
+import mongoose from "mongoose";
+import dotenv from "dotenv";
+import cors from "cors";
 
-// const express = require('express')
-// const dotenv = require("dotenv")
+import bookRoute from "./route/book.route.js";
+import userRoute from "./route/user.route.js";
 
-const app = express()
+const app = express();
+
+app.use(cors());
+app.use(express.json());//parsing in json
+
 dotenv.config();
 
 const PORT = process.env.PORT || 4000;
-const URI  = process.env.MongoDBURI;
+const URI = process.env.MongoDBURI;
 
-//connect to database
+// connect to mongoDB
 try {
-    mongoose.connect(URI,{
+    mongoose.connect(URI, {
         useNewUrlParser: true,
-        useUnifiedTopology: true
-    })
-    console.log("connected to mongo")
+        useUnifiedTopology: true,
+    });
+    console.log("Connected to mongoDB");
 } catch (error) {
-    console.log("Error: ", error)
+    console.log("Error: ", error);
 }
 
-
+// defining routes
+app.use("/book", bookRoute);
+app.use("/user", userRoute);
 
 app.listen(PORT, () => {
-  console.log(`Server is listening on port ${PORT}`)
-})
+    console.log(`Server is listening on port ${PORT}`);
+});
