@@ -1,13 +1,27 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import Slider from "react-slick";
-import list from '../../public/list.json';
 import Cards from './Cards';
+import axios from 'axios';
 function Freebook() {
     //step1: filter free category books
     //react ibrary - slick slider 
-    const filterData = list.filter((data)=> data.category === "Free");//filter is a array method , it filters the data from our array according to the conditions 
+    const[book,setBook] = useState([])
+    useEffect(() => {
+        const getBook  = async() => {
+            try {
+                const res = await axios.get("http://localhost:4001/book");
+                const data = res.data.filter((data)=> data.category === "Free")
+                console.log(data)
+                setBook(data)
+            } catch (error) {
+                console.log(error)
+            }
+        }
+        getBook();
+    },[]);
+    // const filterData = list.filter((data)=> data.category === "Free");//filter is a array method , it filters the data from our array according to the conditions 
     var settings = {
         dots: true,
         infinite: false,
@@ -55,7 +69,7 @@ function Freebook() {
             </div>
             <div>
                 <Slider {...settings}>
-                    {filterData.map((item)=>(
+                    {book.map((item)=>(
                         <Cards item={item} key={item.id}/>
                     ))}
                 </Slider>
