@@ -1,15 +1,41 @@
 import React from 'react'
-import {Link} from 'react-router-dom'
+import {Link, useLocation, useNavigate} from 'react-router-dom'
 import Login from './Login'
-import { useForm } from "react-hook-form" 
-
+import { useForm } from "react-hook-form"; 
+import axios from 'axios';
+import toast from "react-hot-toast";
 function Signup() {
+    const location = useLocation();
+    const navigate = useNavigate();
+    const from = location.state?.from?.pathname || "/";
     const {
         register,
         handleSubmit,
         formState: { errors },
       } = useForm();
-      const onSubmit = (data) => console.log(data)
+      const onSubmit = async (data) => {
+        const userInfo = {
+          fullname: data.fullname,
+          email: data.email,
+          password: data.password,
+        };
+        await axios
+          .post("http://localhost:4001/user/signup", userInfo)
+          .then((res) => {
+            console.log(res.data);
+            if (res.data) {
+              toast.success("Signup Successfully");
+              navigate(from, { replace: true });
+            }
+            localStorage.setItem("Users", JSON.stringify(res.data.user));
+          })
+          .catch((err) => {
+            if (err.response) {
+              console.log(err);
+              toast.error("Error: " + err.response.data.message);
+            }
+          });
+      };
   return (
     <>
         <div className='flex h-screen items-center justify-center'>
@@ -28,10 +54,10 @@ function Signup() {
                                 type='text' 
                                 placeholder='Enter Your Name' 
                                 className='w-80 px-3 py-1 border rounded-md outline-none'
-                                {...register("Name", { required: true })}
+                                {...register("fullname", { required: true })}
                             />
                             <br/>
-                            {errors.email && <span className='text-sm text-red-500 '>This field is required</span>}
+                            {errors.fullname && <span className='text-sm text-red-500 '>This field is required</span>}
                         </div>
                         {/*Email */}
                         <div className="mt-4 space-y-2">
@@ -58,19 +84,6 @@ function Signup() {
                             />
                             <br/>
                             {errors.password && <span className='text-sm text-red-500 '>This field is required</span>}
-                        </div>
-                        {/* PhoneNo */}
-                        <div className="mt-4 space-y-2">
-                            <span>Phone Number: </span>
-                            <br/>
-                            <input 
-                                type='tel' 
-                                placeholder='Enter Your Number' 
-                                className='w-80 px-3 py-1 border rounded-md outline-none'
-                                {...register("tel", { required: true })}
-                            />
-                            <br/>
-                            {errors.tel && <span className='text-sm text-red-500 '>This field is required</span>}
                         </div>
                         {/* Button */}
                         <div className="flex justify-around mt-4">
